[refactor] project(s) 서비스 리팩토링
- 프로젝트 관련 서비스가 나누어져 있어서 혼동의 여지가 있음
- 왼쪽 Projects 탭에 대응되도록 하나의 파일로 병합
