+++ src/service/index.ts
@@ -2,7 +2,6 @@ import axios from '../utils/apiAxios';
 import issueService from './issueService';
 import statsService from './statsService';
 import projectsService from './projectsService';
-import projectService from './projectService';
 import userService from './userService';
 import visitsService from './visitsService';
 import analysisService from './analysisService';
@@ -13,7 +12,6 @@ export default (() => {
     ...statsService(axios),
     ...projectsService(axios),
     ...userService(axios),
-    ...projectService(axios),
     ...visitsService(axios),
     ...analysisService(axios),
   };
